# api/openapi.yaml
openapi: 3.1.0
info:
  title: pglens HTTP API
  version: 1.0.0
  license:
    name: Apache-2.0
servers:
  - url: /
    description: the pglens server itself
...
security:
  - uiSession: []
  - agentBearer: []
components:
  securitySchemes:
    agentBearer:
      type: http
      scheme: bearer
      description: The agent bootstrap token.
    uiSession:
      type: apiKey
      in: cookie
      name: pglens_session
  schemas:
    Error:
      type: object
      required:
        - error
        - detail
      properties:
        error:
          type: string
        detail:
          type: string
...
    ClusterId:
      type: string
      pattern: '^[0-9]+$'
      description: A uint64 rendered as a decimal string because it exceeds the exact integer range of IEEE-754 numbers.
    InstanceId:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp.
...
    Database:
      type: object
      required: [datname, monitored]
      properties:
        datname: { type: string }
        monitored: { type: boolean }
        skip_reason: { type: [string, 'null'] }
...
    ApiMetric:
      type: object
      required: [ts, value, labels]
      properties:
        ts: { $ref: '#/components/schemas/Timestamp' }
        value: { type: number }
        labels: { type: object, additionalProperties: { type: string } }
    LocksResponse:
      type: object
      required: [instance_id, sampled_at, stale, nodes, sessions]
      properties:
        instance_id: { $ref: '#/components/schemas/InstanceId' }
        sampled_at:
          type: [string, 'null']
          format: date-time
          description: Null when no lock tree has ever been sampled, not an empty timestamp.
        stale: { type: boolean }
        nodes: { type: array, items: { type: object, additionalProperties: true } }
        sessions:
          type: array
          items: { type: object, additionalProperties: true }
          description: All client sessions in the latest lock/activity sample.
    ActivityResponse:
      type: object
      required: [instance_id, stale, metrics]
      properties:
        instance_id: { $ref: '#/components/schemas/InstanceId' }
        stale: { type: boolean }
        metrics:
          type: object
          additionalProperties:
            type: array
            items: { $ref: '#/components/schemas/ApiMetric' }
    DatabasesResponse:
      type: object
      required: [instance_id, databases, not_monitored_count]
      properties:
        instance_id: { $ref: '#/components/schemas/InstanceId' }
        databases: { type: array, items: { $ref: '#/components/schemas/Database' } }
        not_monitored_count: { type: integer, format: int32 }
...
    Finding:
      type: object
      required: [finding_id, rule_id, severity, state, scope, title]
      properties:
        finding_id: { type: string }
        rule_id: { type: string }
        severity: { type: string }
        state: { type: string, enum: [open, degraded, muted, resolved] }
        scope: { type: string }
        cluster_id: { $ref: '#/components/schemas/ClusterId' }
        instance_id: { $ref: '#/components/schemas/InstanceId' }
        datname: { type: string }
        object_name: { type: string }
        title: { type: string }
        detail: { type: string }
        remediation: { type: string }
        evidence: { type: object, additionalProperties: true }
        degraded_reason: { type: string }
...
    MuteFindingRequest:
      type: object
      required: [reason, until]
      properties:
        reason: { type: string }
        until: { $ref: '#/components/schemas/Timestamp' }
    MuteFindingResponse:
      type: object
      required: [finding_id, state, muted_until, mute_reason]
      properties:
        finding_id: { type: string }
        state: { type: string, enum: [muted] }
        muted_until: { $ref: '#/components/schemas/Timestamp' }
        mute_reason: { type: string }
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/v1/instances/{id}/activity:
    get:
      tags: [instances]
      operationId: getInstanceActivity
      summary: Return recent activity metrics grouped by metric name.
      parameters:
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/InstanceId' } }
      responses:
        '200':
          description: Recent activity; stale true means no fresh sample was observed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivityResponse' }
              examples:
                default:
                  value:
                    instance_id: 11111111-1111-1111-1111-111111111111
                    stale: false
                    metrics:
                      pg_connections_by_database:
                        - { ts: '2026-08-27T02:00:00Z', value: 3, labels: { datname: app } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/instances/{id}/databases:
    get:
      tags: [instances]
      operationId: getInstanceDatabases
      summary: List databases and monitoring status for an instance.
      parameters:
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/InstanceId' } }
      responses:
        '200':
          description: Database inventory.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatabasesResponse' }
              examples:
                default:
                  value:
                    instance_id: 11111111-1111-1111-1111-111111111111
                    databases: [{ datname: app, monitored: true, skip_reason: null }]
                    not_monitored_count: 0
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/v1/locks:
    get:
      tags: [instances]
      operationId: getLocks
      summary: Return the latest sampled blocking tree.
      parameters:
        - name: instance_id
          in: query
          required: true
          schema: { $ref: '#/components/schemas/InstanceId' }
      responses:
        '200':
          description: Lock tree; sampled_at null means no tree was ever sampled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LocksResponse' }
              examples:
                default:
                  value:
                    instance_id: 11111111-1111-1111-1111-111111111111
                    sampled_at: null
                    stale: true
                    nodes: []
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/v1/findings:
    get:
      tags: [findings]
      operationId: getFindings
      summary: List advisor findings.
      parameters:
        - { name: state, in: query, required: false, schema: { type: string, default: open,degraded } }
        - { name: severity, in: query, required: false, schema: { type: string } }
        - { name: rule_id, in: query, required: false, schema: { type: string } }
        - { name: datname, in: query, required: false, schema: { type: string } }
        - { name: scope, in: query, required: false, schema: { type: string } }
        - { name: instance_id, in: query, required: false, schema: { $ref: '#/components/schemas/InstanceId' } }
        - { name: cluster_id, in: query, required: false, schema: { $ref: '#/components/schemas/ClusterId' } }
        - name: limit
          in: query
          required: false
          schema: { type: integer, format: int32, default: 100, minimum: 1, maximum: 1000 }
          description: Maximum findings; default 100 and maximum 1000.
      responses:
        '200':
          description: Findings ordered by severity and recency.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Finding' } }
              examples:
                default:
                  value:
                    - finding_id: query.slow_mean/11111111-1111-1111-1111-111111111111
                      rule_id: query.slow_mean
                      severity: warning
                      state: open
                      scope: instance
                      title: High mean query latency
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/findings/{finding-id}:
    get:
      tags: [findings]
      operationId: getFinding
      summary: Return one advisor finding.
      parameters:
        - { name: finding-id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Finding detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Finding' }
              examples:
                default:
                  value:
                    finding_id: query.slow_mean/11111111-1111-1111-1111-111111111111
                    rule_id: query.slow_mean
                    severity: warning
                    state: open
                    scope: instance
                    title: High mean query latency
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/findings/{finding-id}/mute:
    post:
      tags: [findings]
      operationId: muteFinding
      summary: Mute a finding until a future timestamp.
      parameters:
        - { name: finding-id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MuteFindingRequest' }
            examples:
              default:
                value: { reason: maintenance, until: '2026-08-29T11:00:00Z' }
      responses:
        '200':
          description: Finding muted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MuteFindingResponse' }
              examples:
                default:
                  value:
                    finding_id: query.slow_mean/11111111-1111-1111-1111-111111111111
                    state: muted
                    muted_until: '2026-08-29T11:00:00Z'
                    mute_reason: maintenance
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503':
          description: Findings storage is unavailable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [findings]
      operationId: unmuteFinding
      summary: Remove a finding mute.
      parameters:
        - { name: finding-id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Finding unmuted; no response body. }
        '404': { $ref: '#/components/responses/NotFound' }
        '503':
          description: Findings storage is unavailable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...